+++ changes.md
@@ -3,7 +3,7 @@
 ![1.1.0](https://img.shields.io/badge/1.1.0-in_development-orange.svg?style=flat-square)
 
 ![1.0.1](https://img.shields.io/badge/1.0.1-in_development-orange.svg?style=flat-square)
-- Upgraded to Graphene 1.1.2, fixes Lua 5.1 module loading
+- Upgraded to Graphene 1.1.3, fixes several issues
 - Added missing `Carbon.Unpack` method for faster unpacking
 - Added alias to `Graphene.Support` as `Carbon.Support`
 - Added `SyntaxErrorException` for Carbide