+++ changes.md
@@ -2,7 +2,7 @@
 
 ![1.1.0](https://img.shields.io/badge/1.1.0-in_development-orange.svg?style=flat-square)
 
-![1.0.1](https://img.shields.io/badge/1.0.1-in_development-orange.svg?style=flat-square)
+![1.0.1](https://img.shields.io/badge/1.0.1-latest-brightgreen.svg?style=flat-square)
 - Upgraded to Graphene 1.1.3, fixes several issues
 - Added missing `Carbon.Unpack` method for faster unpacking
 - Added alias to `Graphene.Support` as `Carbon.Support`
@@ -15,5 +15,5 @@
 - Improved documentation on matrix methods
 - Improved test coverage
 
-![1.0.0](https://img.shields.io/badge/1.0.0-latest-brightgreen.svg?style=flat-square)
+![1.0.0](https://img.shields.io/badge/1.0.0-unsupported-red.svg?style=flat-square)
 - Initial release